requirements.txt
discord.py
requests
ephem